Pagination: Create Pipe + move type to own folder
