+++ src/file/file.service.ts
@@ -4,7 +4,7 @@ import { FileNotFoundFromIDException, FileNotFoundFromPathException, FileNotFoun
 import PrismaService from 'src/prisma/prisma.service';
 import type { Library, File } from '@prisma/client';
 import FileQueryParameters from './models/file.query-parameters';
-import { buildPaginationParameters as buildPaginationParameters, PaginationParameters } from 'src/utils/pagination';
+import { buildPaginationParameters, PaginationParameters } from 'src/pagination/parameters';
 import { FileNotReadableException } from 'src/file-manager/file-manager.exceptions';
 
 @Injectable()
